[INTERNAL,TEST] Separate filesystem configuration and temporary-file policy
Move settings methods into FileConfig.cpp and temporary-directory/registry
implementation into FileTemp.cpp. Keep public definitions and method bodies
unchanged, with core File.cpp free of settings and temp-policy references.

Part of OpenMS/OpenMS#10083, step 5; follows step 4.
